Update index.html
Added the ICRA requirement for extended abstracts: Workshop papers should encourage and prioritize the submission of new ideas. A workshop paper submission may be a non-archival abstract based on a prior publication. However, it cannot take the form of an existing paper that is published or under review, and it cannot be considered published as a peer-reviewed paper.